+++ apps/ddaas/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @windrun-huaiin/ddaas-website
 
+## 14.2.0
+
+### Minor Changes
+
+- feat(mdx): add loading animate component
+
+### Patch Changes
+
+- Updated dependencies
+  - @windrun-huaiin/base-ui@14.0.3
+  - @windrun-huaiin/third-ui@14.3.0
+  - @windrun-huaiin/lib@14.0.1
+
 ## 14.1.2
 
 ### Patch Changes